alexa-skill/requirements.txt
Flask-SQLAlchemy==2.0
SQLAlchemy==1.0.8
